Network handles were being truncated from 64-bit (`long`) to 32-bit (`int`), causing both WiFi and Cellular connections to receive the same invalid handle `-889270259`
Virtual IP Benefits for Network Binding
The current approach uses real interface IPs (172.20.10.2, 192.0.0.2), but virtual IPs could provide:

1. Network Isolation
Each connection gets a dedicated virtual IP in separate subnets
Example: WiFi=10.0.1.1, Cellular=10.0.2.1
Prevents routing conflicts between interfaces
2. Simplified Routing
Virtual IPs can be configured with specific routing tables
Forces traffic through intended network interfaces
Reduces dependency on Android's network binding
3. Better Error Handling
"Machine is not on the network" error suggests routing issues
Virtual IPs with custom routes might bypass this
4. SRTLA Server Compatibility
Some SRTLA servers work better with predictable IP ranges
Virtual IPs can be in controlled subnets
🤔 But Let's First Test the Network Handle Fix
The current network handle fix should resolve the main issue. The truncation from 64-bit to 32-bit was definitely causing problems.

Expected improvements:

✅ Different network handles for WiFi vs Cellular
✅ Proper android_setsocknetwork() calls with correct handles
✅ Better socket binding success rate
Would you like to test the current fix first, or should we implement virtual IPs as an additional enhancement? The network handle fix might be sufficient to resolve the binding issues.

+++ src/main/cpp/srtla_android_wrapper.cpp
@@ -26,7 +26,7 @@ extern "C" {
 struct srtla_android_connection {
     int fd;
     std::string local_ip;
-    int network_handle;
+    long network_handle;
     struct sockaddr_in local_addr;
     struct sockaddr_in server_addr;
     time_t last_sent;
@@ -148,7 +148,7 @@ void SRTLAAndroidWrapper::shutdown() {
     LOGI("SRTLA session shut down");
 }
 
-bool SRTLAAndroidWrapper::addConnection(const std::string& local_ip, int network_handle) {
+bool SRTLAAndroidWrapper::addConnection(const std::string& local_ip, long network_handle) {
     if (!running_) {
         LOGE("Session not running");
         return false;
@@ -191,11 +191,11 @@ bool SRTLAAndroidWrapper::addConnection(const std::string& local_ip, int network
     if (network_handle != 0) {
         net_handle_t net_handle = static_cast<net_handle_t>(network_handle);
         if (android_setsocknetwork(net_handle, connection->fd) != 0) {
-            LOGD("Warning: Could not bind to network handle %d for %s: %s", 
+            LOGD("Warning: Could not bind to network handle %ld for %s: %s", 
                  network_handle, local_ip.c_str(), strerror(errno));
             // Continue anyway - socket can still work without network binding
         } else {
-            LOGD("Successfully bound socket to network handle %d for %s", 
+            LOGD("Successfully bound socket to network handle %ld for %s", 
                  network_handle, local_ip.c_str());
         }
     }
@@ -213,7 +213,7 @@ bool SRTLAAndroidWrapper::addConnection(const std::string& local_ip, int network
     
     session_->connections.push_back(std::move(connection));
     
-    LOGI("Added SRTLA connection: %s (network_handle=%d)", 
+    LOGI("Added SRTLA connection: %s (network_handle=%ld)", 
          local_ip.c_str(), network_handle);
     
     return true;
@@ -328,7 +328,7 @@ Java_com_example_srtla_SRTLANative_shutdown(JNIEnv *env, jobject thiz, jlong ses
 
 JNIEXPORT jboolean JNICALL
 Java_com_example_srtla_SRTLANative_addConnection(JNIEnv *env, jobject thiz, jlong session_ptr,
-                                                 jstring local_ip, jint network_handle) {
+                                                 jstring local_ip, jlong network_handle) {
     auto* wrapper = reinterpret_cast<SRTLAAndroidWrapper*>(session_ptr);
     if (!wrapper) return JNI_FALSE;
     
@@ -336,7 +336,7 @@ Java_com_example_srtla_SRTLANative_addConnection(JNIEnv *env, jobject thiz, jlon
     std::string ip(ip_cstr);
     env->ReleaseStringUTFChars(local_ip, ip_cstr);
     
-    return wrapper->addConnection(ip, network_handle) ? JNI_TRUE : JNI_FALSE;
+    return wrapper->addConnection(ip, (long)network_handle) ? JNI_TRUE : JNI_FALSE;
 }
 
 JNIEXPORT void JNICALL
